In Memory DB Added With Products

+++ CheckoutKata/ICheckout.cs
@@ -3,5 +3,6 @@ namespace CheckoutKata;
 public interface ICheckout
 {
     public void Scan(string item);
-    public int GetTotalPrice();
+    
+    public double GetTotalPrice();
 }
